fix: 调整日志时间选取的交互
