+++ src/panels/SmartLocksTwo/Common/Log/Log.less
@@ -1,4 +1,8 @@
 @import url('./LogList.less');
 .log {
-
+  .date-pick {
+    position: fixed;
+    bottom: 44px;
+    right: 16px;
+  }
 }
