+++ src/panels/SmartLocksTwo/Common/Log/LogList.less
@@ -10,10 +10,4 @@
   .list {
     padding-bottom: 16px;
   }
-  .date-pick {
-    position: fixed;
-    bottom: 44px;
-    right: 16px;
-  }
-
 }
